Additional functionality, migration file updated

+++ EFCoreApplication/Models/TaskModel.cs
@@ -10,7 +10,7 @@ namespace EFCoreApplication.Models
     public string Text { get; set; }
     public DateTime StartDate { get; set; }
     public int Duration { get; set; }
-    public decimal Progress { get; set; }
+    public double Progress { get; set; }
     public int? ParentId { get; set; }
     public string Type { get; set; }
